Separate CommentBox tests into separate file.
TODO: Custom test environment for Puppeteer where the browser instance doesn't
shutdown after each test.
