+++ package.json
@@ -19,7 +19,7 @@
         "stats": "webpack --profile --json > stats.json",
         "publish-ext": "node publish",
         "test": "jest --verbose --silent --config=.jest-config.json",
-        "test:watch": "jest --verbose --watch --config=.jest-config.json",
+        "test:watch": "jest --runInBand --verbose --watch --config=.jest-config.json",
         "test:coverage": "jest --silent --coverage --config=.jest-config.json",
         "lint": "npm-run-all --parallel lint:*",
         "lint:ts": "tslint --project . --format stylish",
